WV-59: подправлено поле в тесте

+++ src/internal/app/track/test/delivery/http/track_handler_local_test.go
@@ -49,7 +49,7 @@ type TestDomainCreator struct{}
 func (creator TestDomainCreator) PrepareOneTestDomain() utilsInterfaces.Domain {
 	return domain.Track{
 		Id:             7,
-		AlbumId:        5,
+		AlbumId:        uint64(5),
 		ArtistId:       3,
 		Title:          "testTrack",
 		Duration:       300,
